+++ lib/get_localization.dart
@@ -1,3 +1,5 @@
 library get_localization;
 
 export 'src/base_localization.dart';
+export 'src/extensions.dart';
+export 'src/localization_mixin.dart';